Task 194 asked for a "Now" button in the javascript datepicker; this was primarily for use in NCS Navigator where many surveys asked for a date/time stamp to monitor the process for the current survey.
In discussions about surveyor, we decided to move from the default jquery tools date input to the jquery ui datepicker for date, time, and datetime fields.

This commit addresses that change, plus notes on how to configure surveyor to use either datepicker or date input - or to not use javascript at all for dates and rely on the rails date helpers (multiple dropdown lists for each date time part).