Loosen Rake dependency and make it development-only.
We cannot use Rake 0.8.7; see #281.  Rake 0.9.2 is known to work.  I've also
tried Rake 0.9.2.2, and that seems to work, so I think it's appropriate to
loosen the version constraint.

I heard from @rsutphin that Jim Weirich mentioned that he'd be changing his
versioning schema to drop the leading zero from Rake versions, so that the next
release of Rake will be 10.x.  In light of that information -- and assuming
Rake keeps its rate of introducing regressions low, which in my experience it
does -- I think a >= is appropriate.

Loosing the constraint also solves the problem encountered in #309.  This patch
is more-or-less the same solution that @gaurish devised.