Move tmpdir for CI. #333.
For some reason, when $TMPDIR is set to `pwd`/tmp (i.e., in the build
workspace), the gconfd-2 process started by Firefox during the Selenium-backed
tests stops responding (causing the test to fail) and eventually eats all the
memory on the machine (causing Jenkins to die). Using the system default
TMPDIR (/tmp in our case) does not show this problem, but makes it so that
builds share temporary files and so might interfere with each other. Moving
$TMPDIR to a unique-per-build-params directory under $HOME resolves this
issue.