Render grid hidden inputs in a table cell. Closes #339.
One of the effects of the changes made for the upgrade to Formtastic 2.1
(#227) apparently had the effect of wrapping the hidden inputs for question
ID and response ID in `li` tags. Those inputs were being rendered directly
into a `tr` (i.e., outside of any cell). This was illegal nesting, so the
browser relocated the `li`s (and the inputs) to somewhere they could legally
be rendered.

This seems to have moved them enough that they were no longer matched by the
jQuery selector used for submitting partial results. They were still in the
`form` overall and so were submitted when the whole page was submitted.

Moving the inputs to somewhere an `li` can legally be nested fixes this
problem.