Features for ajax response submissions. Done with #328.
This is a longstanding Surveyor behavior, but the existing features (including the
selenium ones) mostly did not exercise it.

Some of the features are `@wip`, either because they are incomplete (I can't figure out
a good way to accurately simulate a user's use of a jQuery UI slider) or because they create duplicates per #328 and so fail.