Consolidated update-from-ui method. #328.
This method relies on receiving api_ids for both new and existing responses to
eliminate the chance of duplicates. Still incomplete here: demonstrating what
exception is thrown when there's a race between to different attempts to
create the same response. Still incomplete everywhere else:
generating/rendering api_ids for responses in the UI.