Change test infra so that tx behavior can be tested. Fixes #360.
ActiveRecord 3.0 does not roll back nested transactions with SQLite.
The default rails (and spec-rails) mode of execution wraps every
test in a transaction, so a transaction in the code under test
becomes a nested one.

This commit works around this by diabling the default rails test
behavior and switching to DatabaseCleaner. It still uses
transactions for most of the tests, but allows you to switch to
truncation for a particular example or context.