ResponseSet.new shouldn't require database access.
Additionally, ResponseSet#random_survey_access_code does a lot of work
for no reason.  Repeated searches for an existing access code does not
guarantee that the access code is unique, especially in the common case
when multiple writers are present.

The existing code only protects against the case in which the generated
code was used in a previously committed dataset.  That case can be made
sufficiently improbable by minimizing regularity in the access code
generator.

Ruby 1.9 contains a class SecureRandom which (is supposed to) implement
a cryptographically secure PRNG.  Therefore, assuming SecureRandom holds
up its end of the contract, the values it generates will be
statistically random, which means something like this can be used to
generate survey access codes:

    module Surveyor::Common
      def self.make_tiny_code
        # Base64 encoding increases size by ~4/3;
        # ceil(7 * 4/3) = 10
        SecureRandom.urlsafe_base64(7)
      end
    end

More about this in a future commit.