Use SecureRandom for generating access codes.
This commit contains a fallback for Ruby < 1.9 via
ActiveSupport::SecureRandom, which exists in ActiveSupport 2.2.1 through
3.0.9.

How good is it?  Or perhaps more precisely:

    What is the probability that two or more response sets out of a
    group of n response sets will have the same access code?

This is an instance of the birthday problem.

We have 256^7 possible access codes.  (Each access code is a 7-byte
string; we assume that each byte of the string is equally likely to come
from the set of all possible bytes.  SecureRandom should make that
true.)

Let P(n, d) be the probability that two or more response sets from a
group of n response sets have the same access code (out of d possible
access codes).

    P(n, d) = 1 - (d! / ((d - n)! * d^n))

This can be simplified to remove factorials:

    d! / ((d - n)! * d^n)
    = (d! / (d - n)!) * (1/d^n)
    = product(i = 0..n - 1, d - i) * (1/d^n)
    = product(i = 0..n - 1, d - i) / d^n

    => P(n, d) = 1 - product(i = 0..n - 1, d - i) / d^n

For various values of n, we get

    |   n   | P(n, 256^7)  |
    | 10    | 6.3 * 10^-16 |
    | 100   | 6.9 * 10^-14 |
    | 1000  | 6.9 * 10^-12 |
    | 10000 | 6.9 * 10^-10 |

That is, with 10,000 response sets, the chance of a duplicate access
code is around one in 1.5 billion.  IMO, this is good enough.

As a point of comparison, v4 UUIDs (122 random bits vs. our 56) have
these probabilities:

    |   n   | P(n, 2^122)  |
    | 10    | 8.5 * 10^-36 |
    | 100   | 9.3 * 10^-34 |
    | 1000  | 9.4 * 10^-32 |
    | 10000 | 9.4 * 10^-30 |

Finally, in the case of collision: the old code still had a case where
it wouldn't gracefully handle collisions.  To gracefully handle an
access code collision, Surveyor needs code that can generate a new
access code and restart a save transaction.