SecureRandom was backported to 1.8.7, so we can use it.  #370.
We still need to do feature detection for .urlsafe_base64 (it was
introduced in the 1.9 series), but that's not too bad.

I think there's a benefit in keeping around two implementations with an
eye towards using only SecureRandom.urlsafe_base64; the simpler
implementation more precisely describes what we want.  However, we can
eliminate the conditional by just adopting the version that doesn't use
.urlsafe_base64.