Move access code and version generation to callbacks.
These callbacks also make an allocation for explicitly-set access codes
by handling them as one would expect, i.e. not overwriting them because
it's not necessary to do so.

Reasons:

1. This exhibits the same behavior as far as Surveyor's examples are
   concerned, so no spec changes were necessary.  On my machine,
   Surveyor's full test suite passes with this change.
2. ActiveRecord models are expected to not engage in database activity
   from attribute accessors, unless e.g. said accessors are association
   proxies with autosave enabled.  However, #title= does not fall into
   that category.
3. NCS Navigator Cases' spec suite exhibited strange behaviors in which
   the overridden #title= was not executed when building valid (up to
   syntax and reference checks) test surveys via Surveyor's parser.  I
   cannot replicate the problem in a Surveyor testbed, but due to reason
   #2 I decided it wasn't worth the time because the existing code was
   already pretty janky.
4. This approach is more efficient.  If Survey#title= is invoked
   multiple times on the same object, one query will be generated per
   invocation.  With this approach, only one such query will be issued.
   (The generated query is also more efficient.)

Of course, one shouldn't be generating queries for incrementing versions
at all.  The approach has the same fundamental deficiency as unlocked
existence tests for uniqueness, and the same real solutions, e.g. put
the check where you can actually enforce an ordering on operations or --
even better -- don't rely on a separate counter.  But that's a battle
for another commit.