locale selection in survey UI #427
The locale is modified by submitting the survey via javascript
onchange of a select box. The feature is not visible when javascript
is disabled. The current section is kept when changing languages and vise versa.

A work-around for #426 is provided here, which uses I18n fallbacks and
'default' locale if the default locale of a survey is not provided.